Fix performance
splitting home component, creating SearchForm.

Applying react memo to gif.
