+++ src/components/Gif.jsx
@@ -13,4 +13,4 @@ function Gif({ title, id, url }) {
   );
 }
 
-export default Gif;
+export default React.memo(Gif);
