+++ src/pages/Home/index.jsx
@@ -1,35 +1,24 @@
-import React, { useState } from "react";
+import React, { useCallback } from "react";
 import { useLocation } from "wouter";
 import ListOfGifs from "components/ListOfGifs";
 import useGifs from "hooks/useGifs";
 import TrendingSearches from "components/TrendingSearches";
+import SearchForm from "components/SearchForm";
 
 function Home() {
-  const [keyword, setKeyword] = useState("");
   const [path, pushLocation] = useLocation();
   const { gifs, loading } = useGifs();
 
-  const handleSubmit = (event) => {
-    event.preventDefault();
-    //navegar a otra router
-    pushLocation(`/search/${keyword}`);
-  };
-  const handleChange = (event) => {
-    setKeyword(event.target.value);
-  };
+  const handleSubmit = useCallback(
+    ({ keyword }) => {
+      pushLocation(`/search/${keyword}`);
+    },
+    [pushLocation]
+  );
 
   return (
     <>
-      <form onSubmit={handleSubmit}>
-        <input
-          placeholder="Search a gif here..."
-          onChange={handleChange}
-          type="text"
-          value={keyword}
-        />
-        <button type="submit">search</button>
-      </form>
-
+      <SearchForm onSubmit={handleSubmit} />
       <h3>Ultima busqueda</h3>
       <ListOfGifs gifs={gifs} loading={loading} />
